% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pig_latin.R
\name{pig_latin}
\alias{pig_latin}
\title{Translate using pig latin}
\usage{
pig_latin(word, end = c("ay", "way", "yay"))
}
\arguments{
\item{word}{A character vector of word(s) to translate.}

\item{end}{The ending of words beginning with a vowel.}
}
\description{
Translate words into their pig latin form.
}
\examples{
pig_latin("hello")
pig_latin("airplane")
pig_latin("eye")
pig_latin("ear")
pig_latin("aeiou")
pig_latin("my")
pig_latin(c("hello", "world"))

}
